[UPDATE] Conversão fr 'quantity' mudou para float.

+++ l10n_br_account_product/account_invoice.py
@@ -999,7 +999,7 @@ class AccountInvoiceLine(orm.Model):
         price = float(price_unit) * (1 - values.get('discount', 0.0) / 100.0)
 
         taxes_calculed = tax_obj.compute_all(
-            cr, uid, taxes, price, int(values.get('quantity', 0.0)),
+            cr, uid, taxes, price, float(values.get('quantity', 0.0)),
             values.get('product_id'), partner_id,
             fiscal_position=fiscal_position,
             insurance_value=values.get('insurance_value', 0.0),
